+++ source/extensions/omni.flux.asset_importer.widget/omni/flux/asset_importer/widget/tests/e2e/file_import_list/test_file_import_list_widget.py
@@ -462,7 +462,7 @@ class TestFileImportListWidget(omni.kit.test.AsyncTestCase):
         self.assertEqual(len(file_items), 0)
 
         # Simulate dropping files on the widget
-        widget._on_drag_drop_external(event)
+        widget.on_drag_drop_external(event)
         await ui_test.human_delay()
 
         file_items = ui_test.find_all(f"{window.title}//Frame/**/Label[*].identifier=='file_path'")
@@ -494,7 +494,7 @@ class TestFileImportListWidget(omni.kit.test.AsyncTestCase):
 
         # Simulate dropping files on the widget The failed drop will raise an error dialog that we don't need here.
         with patch("omni.flux.asset_importer.widget.file_import_list.widget._file_validation_failed_callback"):
-            widget._on_drag_drop_external(event)
+            widget.on_drag_drop_external(event)
             await ui_test.human_delay()
 
         file_items = ui_test.find_all(f"{window.title}//Frame/**/Label[*].identifier=='file_path'")
@@ -525,7 +525,7 @@ class TestFileImportListWidget(omni.kit.test.AsyncTestCase):
 
         # Simulate dropping files on the widget The failed drop will raise an error dialog that we don't need here.
         with patch("omni.flux.asset_importer.widget.file_import_list.widget._file_validation_failed_callback"):
-            widget._on_drag_drop_external(event)
+            widget.on_drag_drop_external(event)
             await ui_test.human_delay()
 
         file_items = ui_test.find_all(f"{window.title}//Frame/**/Label[*].identifier=='file_path'")