+++ source/extensions/omni.flux.validator.plugin.selector.usd/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.9.5]
+### Changed
+- Updated internal import paths for prim utilities
+
 ## [1.9.4]
 ### Changed
 - Applied new lint rules